refactor config, tokenRouter
